New: Postgres Support

+++ src/NzbDrone.Core/Datastore/Migration/133_add_minimumavailability.cs
@@ -13,12 +13,12 @@ namespace NzbDrone.Core.Datastore.Migration
         {
             if (!Schema.Schema("dbo").Table("NetImport").Column("MinimumAvailability").Exists())
             {
-                Alter.Table("NetImport").AddColumn("MinimumAvailability").AsInt32().WithDefaultValue(MovieStatusType.Released);
+                Alter.Table("NetImport").AddColumn("MinimumAvailability").AsInt32().WithDefaultValue((int)MovieStatusType.Released);
             }
 
             if (!Schema.Schema("dbo").Table("Movies").Column("MinimumAvailability").Exists())
             {
-                Alter.Table("Movies").AddColumn("MinimumAvailability").AsInt32().WithDefaultValue(MovieStatusType.Released);
+                Alter.Table("Movies").AddColumn("MinimumAvailability").AsInt32().WithDefaultValue((int)MovieStatusType.Released);
             }
         }
     }
